backend/requirements.txt
Flask==3.0.0
Flask-PyMongo==2.3.0
Flask-JWT-Extended==4.6.0
Flask-SocketIO==5.3.6
Flask-CORS==4.0.0
Flask-Mail==0.9.1
bcrypt==4.1.2
cloudinary==1.37.0
razorpay==1.4.1
python-dotenv==1.0.0
gunicorn==21.2.0
eventlet==0.35.1
Werkzeug==3.0.1
pymongo==4.6.1
